requirements.txt
autopep8==1.4.4
boto3==1.9.200
botocore==1.12.200
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.4
django-bootstrap-tags==1.2.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-filter==2.1.0
django-foreign-fields==0.1
django-forms-bootstrap==3.1.0
django-storages==1.7.1
docutils==0.14
future==0.17.1
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
pep8==1.7.1
Pillow==6.0.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
python-dateutil==2.8.0
python-http-client==3.0.0
Python-Rest-Framework==0.2.2
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.12.0
sqlparse==0.3.0
stripe==2.32.1
urllib3==1.25.3
whitenoise==4.1.2
